Debug screen remake
